+++ R/imports.R
@@ -4,7 +4,8 @@
 #' @importFrom jsonlite toJSON
 #' @importFrom digest digest
 #' @importFrom cli cli_abort
-#' @importFrom rlang abort
+#' @importFrom codetools findGlobals
+#' @importFrom rlang abort .data
 #' @importFrom tibble as_tibble
 #' @keywords internal
 NULL